Deleting OrderService and passing its logic to OrdersManagerController
